+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/event/processor/SystemMappingDeleteProcessor.java
@@ -5,11 +5,15 @@ import org.springframework.context.annotation.Description;
 import org.springframework.stereotype.Component;
 import org.springframework.util.Assert;
 
+import com.google.common.collect.ImmutableMap;
+
+import eu.bcvsolutions.idm.acc.domain.AccResultCode;
 import eu.bcvsolutions.idm.acc.dto.SysSystemMappingDto;
 import eu.bcvsolutions.idm.acc.dto.filter.SysRoleSystemFilter;
 import eu.bcvsolutions.idm.acc.dto.filter.SysSystemAttributeMappingFilter;
 import eu.bcvsolutions.idm.acc.event.SystemMappingEvent.SystemMappingEventType;
 import eu.bcvsolutions.idm.acc.service.api.SysRoleSystemService;
+import eu.bcvsolutions.idm.acc.service.api.SysSyncConfigService;
 import eu.bcvsolutions.idm.acc.service.api.SysSystemAttributeMappingService;
 import eu.bcvsolutions.idm.acc.service.api.SysSystemMappingService;
 import eu.bcvsolutions.idm.core.api.event.CoreEvent;
@@ -17,6 +21,7 @@ import eu.bcvsolutions.idm.core.api.event.CoreEventProcessor;
 import eu.bcvsolutions.idm.core.api.event.DefaultEventResult;
 import eu.bcvsolutions.idm.core.api.event.EntityEvent;
 import eu.bcvsolutions.idm.core.api.event.EventResult;
+import eu.bcvsolutions.idm.core.api.exception.ResultCodeException;
 
 /**
  * Processor remove all handled attributes for {@link SysSystemMappingDto}
@@ -34,26 +39,34 @@ public class SystemMappingDeleteProcessor extends CoreEventProcessor<SysSystemMa
 	private final SysSystemAttributeMappingService systemAttributeMappingService;
 	private final SysRoleSystemService roleSystemService;
 	private final SysSystemMappingService systemMappingService;
+	private final SysSyncConfigService configService;
 	
 	@Autowired
 	public SystemMappingDeleteProcessor(
 			SysSystemAttributeMappingService systemAttributeMappingService,
 			SysRoleSystemService roleSystemService,
-			SysSystemMappingService systemMappingService) {
+			SysSystemMappingService systemMappingService,
+			SysSyncConfigService configService) {
 		super(SystemMappingEventType.DELETE);
 		//
 		Assert.notNull(roleSystemService);
 		Assert.notNull(systemAttributeMappingService);
 		Assert.notNull(systemMappingService);
+		Assert.notNull(configService);
 		//
 		this.roleSystemService = roleSystemService;
 		this.systemAttributeMappingService = systemAttributeMappingService;
 		this.systemMappingService = systemMappingService;
+		this.configService = configService;
 	}
 
 	@Override
 	public EventResult<SysSystemMappingDto> process(EntityEvent<SysSystemMappingDto> event) {
 		SysSystemMappingDto systemMapping = event.getContent();
+		//
+		if (configService.countBySystemMapping(systemMapping) > 0) {
+			throw new ResultCodeException(AccResultCode.SYSTEM_MAPPING_DELETE_FAILED_USED_IN_SYNC, ImmutableMap.of("mapping", systemMapping.getName()));
+		}
 		//
 		// remove all handled attributes
 		SysSystemAttributeMappingFilter filter = new SysSystemAttributeMappingFilter();